refactor: spring cleaning
