+++ containers/ecr-viewer/integration/conditions.test.ts
@@ -2,7 +2,7 @@
  * @jest-environment node
  */
 
-import { getAllConditions } from "@/app/data/conditions";
+import { getAllConditions } from "@/app/services/listConditionsService";
 
 import { createEcrCondition } from "./helpers/core";
 import { buildCore, dropExisting } from "./helpers/ddl";
