+++ containers/ecr-viewer/integration/saveFhirDataService/core.test.ts
@@ -3,11 +3,11 @@
  */
 
 import { buildCore, clearCore, dropExisting } from "../helpers/ddl";
-import { saveFhirMetadata } from "@/app/api/save-fhir-data/save-fhir-data-service";
+import { saveFhirMetadata } from "@/app/api/save-fhir-data/service";
 import { BundleMetadata } from "@/app/api/save-fhir-data/types";
-import { getDb } from "@/app/api/services/database";
-import { Core } from "@/app/api/services/types/core";
 import { BlobResponse } from "@/app/data/blobStorage/utils";
+import { getDb } from "@/app/data/metadataDb/database";
+import { Core } from "@/app/data/metadataDb/types/core";
 
 const baseCoreMetadata: BundleMetadata = {
   last_name: "lname",
