+++ containers/ecr-viewer/src/app/api/fhir-data/service.ts
@@ -4,15 +4,15 @@ import { ApiError } from "@google-cloud/storage";
 import { Bundle } from "fhir/r4";
 import { NextResponse } from "next/server";
 
+import { azureBlobContainerClient } from "@/app/data/blobStorage/azureClient";
+import { gcpClient } from "@/app/data/blobStorage/gcpClient";
+import { s3Client } from "@/app/data/blobStorage/s3Client";
 import {
   AZURE_SOURCE,
   GCP_SOURCE,
   S3_SOURCE,
   streamToJson,
-} from "@/app/api/utils";
-import { azureBlobContainerClient } from "@/app/data/blobStorage/azureClient";
-import { gcpClient } from "@/app/data/blobStorage/gcpClient";
-import { s3Client } from "@/app/data/blobStorage/s3Client";
+} from "@/app/data/blobStorage/utils";
 
 const UNKNOWN_ECR_ID = "eCR ID not found";
