+++ containers/ecr-viewer/src/app/data/blobStorage/gcpClient.ts
@@ -1,7 +1,6 @@
 import { Storage } from "@google-cloud/storage";
 
-import { GCP_SOURCE } from "@/app/api/utils";
-
+import { GCP_SOURCE } from "./utils";
 import {
   BlobResponse,
   DELETE_FAILURE,
