+++ containers/ecr-viewer/src/app/data/metadataDb/types/core.ts
