+++ containers/ecr-viewer/src/app/data/metadataDb/utils/db-config.ts
