+++ containers/ecr-viewer/src/app/tests/api/process-zip/service.test.ts
@@ -6,8 +6,8 @@ import { processZip } from "@/app/api/process-zip/service";
 import {
   saveFhirData,
   saveWithMetadata,
-} from "@/app/api/save-fhir-data/save-fhir-data-service";
-import { S3_SOURCE } from "@/app/api/utils";
+} from "@/app/api/save-fhir-data/service";
+import { S3_SOURCE } from "@/app/data/blobStorage/utils";
 
 jest.mock("../../../api/save-fhir-data/save-fhir-data-service");
 jest.mock("../../../api/services/database");
